BugFix: Crash when population with xyz position grows.
This bug was introduced by 3966b66.

When copying data to tex buffers, the buffer len (of the first buffer, aka position), is checked against the requested agent copy count. As float3 buffers are not supported, a float1 buffer of Nx length is used to support array variables, hence the comparison assumed the buffer was Nx longer than it was.

This fix simply divides the buffer len by that multiplier.