BugFix for Linux screenshots not including occluded parts of backbuffer.
Scene is now rendered to a secondary (multisample) framebuffer, this is then blit to the backbuffer.
When a user requests a screenshot, this is then blit to the screenshot buffer (as glReadPixels() does not support multisample framebuffers).
Also some minor patches to the FrameBuffer code, e.g. making redundant resizes cheaper.